+++ plone/app/layout/analytics/tests/base.py
@@ -1,10 +1,9 @@
-from Products.PloneTestCase import PloneTestCase as ptc
-ptc.setupPloneSite()
+from plone.app.testing.bbb import PloneTestCase
 
 
-class AnalyticsTestCase(ptc.PloneTestCase):
-    pass
+class AnalyticsTestCase(PloneTestCase):
+    """ Base testing class for analytics tests """
 
 
-class AnalyticsFunctionalTestCase(ptc.FunctionalTestCase):
-    pass
+class AnalyticsFunctionalTestCase(PloneTestCase):
+    """ Base testing class for analytics functional tests """